// src/GraphQLAST.d.ts
export type Location = {
    /**
   * The character offset at which this Node begins.
   */
    start: number;

    /**
   * The character offset at which this Node ends.
   */
    end: number;

    /**
   * The Token at which this Node begins.
   */
    startToken: Token;

    /**
   * The Token at which this Node ends.
   */
    endToken: Token;

    /**
   * The Source document the AST represents.
   */
    source: any;
};
// ...
export type Token = {
    /**
   * The kind of Token.
   */
    kind: TokenKind;

    /**
   * The character offset at which this Node begins.
   */
    start: number;

    /**
   * The character offset at which this Node ends.
   */
    end: number;

    /**
   * The 1-indexed line number on which this Token appears.
   */
    line: number;

    /**
   * The 1-indexed column number at which this Token begins.
   */
    column: number;

    /**
   * For non-punctuation tokens, represents the interpreted value of the token.
   */
    value: string | void;

    /**
   * Tokens exist as nodes in a double-linked-list amongst all tokens
   * including ignored tokens. <SOF> is always the first node and <EOF>
   * the last.
   */
    prev: Token | null;
    next: Token | null;
};
// ...
export type NameNode = {
    kind: "Name";
    loc?: Location;
    value: string;
};

// Document

export type DocumentNode = {
    kind: "Document";
    loc?: Location;
    definitions: Array<DefinitionNode>;
};

export type DefinitionNode =
    | OperationDefinitionNode
    | FragmentDefinitionNode
    | TypeSystemDefinitionNode; // experimental non-spec addition.

export type OperationDefinitionNode = {
    kind: "OperationDefinition";
    loc?: Location;
    operation: OperationTypeNode;
    name?: NameNode;
    variableDefinitions?: Array<VariableDefinitionNode>;
    directives?: Array<DirectiveNode>;
    selectionSet: SelectionSetNode;
};

// Note: subscription is an experimental non-spec addition.
export type OperationTypeNode = "query" | "mutation" | "subscription";

export type VariableDefinitionNode = {
    kind: "VariableDefinition";
    loc?: Location;
    variable: VariableNode;
    type: TypeNode;
    defaultValue?: ValueNode;
};

export type VariableNode = {
    kind: "Variable";
    loc?: Location;
    name: NameNode;
};

export type SelectionSetNode = {
    kind: "SelectionSet";
    loc?: Location;
    selections: Array<SelectionNode>;
};

export type SelectionNode = FieldNode | FragmentSpreadNode | InlineFragmentNode;

export type FieldNode = {
    kind: "Field";
    loc?: Location;
    alias?: NameNode;
    name: NameNode;
    arguments?: Array<ArgumentNode>;
    directives?: Array<DirectiveNode>;
    selectionSet?: SelectionSetNode;
};

export type ArgumentNode = {
    kind: "Argument";
    loc?: Location;
    name: NameNode;
    value: ValueNode;
};

// Fragments

export type FragmentSpreadNode = {
    kind: "FragmentSpread";
    loc?: Location;
    name: NameNode;
    directives?: Array<DirectiveNode>;
};

export type InlineFragmentNode = {
    kind: "InlineFragment";
    loc?: Location;
    typeCondition?: NamedTypeNode;
    directives?: Array<DirectiveNode>;
    selectionSet: SelectionSetNode;
};

export type FragmentDefinitionNode = {
    kind: "FragmentDefinition";
    loc?: Location;
    name: NameNode;
    typeCondition: NamedTypeNode;
    directives?: Array<DirectiveNode>;
    selectionSet: SelectionSetNode;
};

// Values

export type ValueNode =
    | VariableNode
    | IntValueNode
    | FloatValueNode
    | StringValueNode
    | BooleanValueNode
    | NullValueNode
    | EnumValueNode
    | ListValueNode
    | ObjectValueNode;
// ...
export type DirectiveNode = {
    kind: "Directive";
    loc?: Location;
    name: NameNode;
    arguments?: Array<ArgumentNode>;
};

// Type Reference

export type TypeNode = NamedTypeNode | ListTypeNode | NonNullTypeNode;

export type NamedTypeNode = {
    kind: "NamedType";
    loc?: Location;
    name: NameNode;
};

export type ListTypeNode = {
    kind: "ListType";
    loc?: Location;
    type: TypeNode;
};

export type NonNullTypeNode = {
    kind: "NonNullType";
    loc?: Location;
    type: NamedTypeNode | ListTypeNode;
};

// Type System Definition

export type TypeSystemDefinitionNode =
    | SchemaDefinitionNode
    | TypeDefinitionNode
    | TypeExtensionDefinitionNode
    | DirectiveDefinitionNode;